backend_agent/pyproject.toml
[project]
name = "backend-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.14.1",
]
